// src/app/models/match.model.ts
export class Match {
  id?: any;
  date?: string;
  result?: string;
  homeTeam?: string;
  visitTeam?: string;
  visitPointGuard?: string;
  visitShootingGuard?: string;
  visitSmallForward?: string;
  visitPowerForward?: string;
  visitCenter?: string;
  homePointGuard?: string;
  homeShootingGuard?: string;
  homeSmallForward?: string;
  homePowerForward?: string;
  homeCenter?: string;
}